refactoring home and user post page, created post component, fixing fetch problem due server side, fixing search component and pagination component
