+++ src/components/Posts/page.module.css
@@ -1,0 +1,225 @@
+
+.filterBar {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 10px;
+  margin-bottom: 20px;
+  /* padding: 10px; */
+  background-color: var(--bg-color);
+  border-radius: 8px;
+  position: relative;
+  margin-top: 40px;
+}
+
+
+.filterBar .filterButton {
+  background-color: var(--post-bg-color);
+  color: var(--text-color);
+  border: 1px solid var(--text-color);
+  padding: 8px 16px;
+  border-radius: 5px;
+  cursor: pointer;
+  transition: background-color 0.3s ease, color 0.3s ease;
+}
+
+
+.filterBar .filterButton:hover {
+  background-color: rgba(0, 0, 0, 0.1);
+}
+
+
+.filterBar .active {
+  background-color: var(--text-color);
+  color: var(--bg-color);
+}
+
+
+.mainContainer {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(300px, 1fr));
+  gap: 20px;
+  padding: 20px;
+  background-color: var(--bg-color);
+  color: var(--text-color);
+  transition: background-color 0.3s ease, color 0.3s ease;
+  min-height: calc(100vh - 200px); 
+  margin-top: -20px;
+  justify-items: center; 
+}
+
+
+.postsWrapper {
+  min-height: calc(100vh - 200px); 
+  margin: 0 auto;
+  padding: 20px;
+  background-color: var(--bg-color);
+}
+
+
+.postCard {
+  background-color: var(--post-bg-color);
+  padding: 20px;
+  border-radius: 10px;
+  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
+  transition: transform 0.3s ease, background-color 0.3s ease;
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  height: 100%;
+  min-height: 350px; 
+  max-height: 450px; 
+  overflow: hidden; 
+}
+
+.postCard:hover {
+  transform: scale(1.05);
+}
+
+
+.imageContainer {
+  width: 100%;
+  height: 200px; 
+  overflow: hidden;
+}
+
+.image {
+  width: 100%;
+  height: 100%;
+  object-fit: cover; 
+}
+
+
+.content {
+  padding: 15px;
+  flex-grow: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  
+}
+
+.title {
+  font-size: 20px;
+  margin-bottom: 10px;
+  color: var(--text-color);
+  text-align: center;
+}
+
+.desc {
+  font-size: 16px;
+  color: var(--text-color);
+  text-align: justify;
+}
+
+.readingTime {
+  font-size: 14px;
+  color: var(--text-color);
+  text-align: right;
+  margin-top: 10px;
+}
+
+
+.likeContainer {
+  margin-top: 15px;
+  text-align: center;
+  margin-bottom: 10px; 
+}
+
+.likeButton {
+  background-color: transparent;
+  color: var(--text-color);
+  border: none;
+  font-size: 16px;
+  cursor: pointer;
+}
+
+.likeButton:hover {
+  color: #e9e9e9;
+}
+
+
+.createPostButton {
+  position: fixed;
+  top: 50%;
+  right: 20px;
+  transform: translateY(-50%);
+  background-color: transparent;
+  color: #b81111;
+  padding: 15px 20px;
+  border-radius: 50%;
+  font-size: 24px;
+  cursor: pointer;
+  z-index: 1000;
+  transition: background-color 0.3s ease, color 0.3s ease, border 0.3s ease;
+}
+
+.createPostButton:hover {
+  background-color: rgba(179, 182, 180, 0.8);
+  color: rgb(241, 13, 43);
+  border-color: rgb(74, 0, 170);
+}
+
+
+.noPostsMessage {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  min-height: calc(100vh - 200px);
+  text-align: center;
+  padding: 40px;  
+  border-radius: 10px;
+  box-shadow: 0px 4px 8px rgba(0, 0, 0, 0.1);
+  transition: background-color 0.3s ease, color 0.3s ease;
+}
+
+.noPostsMessage h2 {
+  font-size: 2rem;
+  color: var(--text-color);
+  margin-bottom: 20px;
+}
+
+.noPostsMessage p {
+  font-size: 1.2rem;
+  color: var(--text-color);
+  margin-bottom: 30px;
+}
+
+
+.createPostLink {
+  color: #0070f3;
+  text-decoration: underline;
+  font-weight: bold;
+  cursor: pointer;
+}
+
+.createPostLink:hover {
+  color: #0056a3;
+}
+
+
+@media (max-width: 600px) {
+  .mainContainer {
+    grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
+  }
+
+  .createPostButton {
+    right: 10px;
+    top: 60%;
+  }
+
+  .noPostsMessage {
+    padding: 20px;
+    min-height: calc(100vh - 180px);
+  }
+
+  .noPostsMessage h2 {
+    font-size: 1.6rem;
+  }
+
+  .noPostsMessage p {
+    font-size: 1rem;
+  }
+}
